% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births_null.R
\name{births_null}
\alias{births_null}
\title{Template births function that returns a fixed number of births (default zero)
for each aggregation level, and sets their age to zero}
\usage{
births_null(pop, col_aggregation = c("gss_code", "sex"), const = 0)
}
\arguments{
\item{pop}{A data frame containing population data}

\item{col_aggregation}{A string giving the names of columns to which the
output births will be aggregated to. Default \code{c("gss_code","sex")}}

\item{const}{Numeric. Number of births to return per geography. Defaults to
zero, but can be set to any positive number}
}
\value{
A data frame of births with one row for each distinct value of the
  \code{col_aggregation} columns, a column named births with value
  \code{const} and a column named age with value 0.
}
\description{
This is designed to be used as a placeholder births function during model
testing and a template for other births functions, along with its tests.
}
